release: prep v0.15.0 — tight-core refactor
Removes 8 commands from core. Net -5827 LOC.

DELETED ENTIRELY (no plugin successor):
- fledge update (templates re-application — bidirectional sync trap)
- fledge metrics (LOC/churn/test ratio — overlapped by tokei/scc/Insights)

DELETED FROM CORE (future plugins):
- fledge deps  → fledge-plugin-deps
- fledge checks/issues/prs  → fledge-plugin-github
- fledge templates search/publish  → fledge-plugin-templates-remote

STUBBED:
- fledge doctor: stripped 250+ LOC of toolchain probes (rust/node/
  python/swift/...). Core checks fledge config + git + AI provider.
  Tool-specific probes belong in fledge-plugin-doctor-* plugins.

KEPT (the signature pillars):
- ai, ask, review (with v0.15's --with-model multi-model panel),
  work, run, lanes, watch, spec, templates (local: init/create/
  validate/list), release, changelog, plugins, config, introspect,
  completions, doctor

The signature reads cleaner: scaffold (templates), run (run/lanes/
watch), spec (spec), AI (ai/ask/review), ship (work/release/
changelog), extend (plugins/config/introspect/completions/doctor).
21 → 14 commands. 28K → 22K LOC.

Cargo.toml authors set to 0xLeif <leif.algo@pm.me>.

Spec count: 35 → 29 (deleted update/deps/metrics/checks/issues/prs).
Bumps Cargo.toml 0.14.0 → 0.15.0.
