--- specs/issues/tasks.md
@@ -1,14 +1,0 @@
----
-spec: issues.spec.md
----
-
-## Tasks
-
-- [x] Write spec files for issues module
-- [x] Implement `issues` list — filtered, styled issue listing
-- [x] Implement `issues view` — detailed single-issue view
-- [x] Add PR detection and redirect in view
-- [x] Add `--state`, `--label`, `--limit`, `--json` flags
-- [x] Wire up CLI subcommands in main.rs
-- [x] Write unit tests
-- [x] Run verification suite
