--- specs/metrics/tasks.md
@@ -1,15 +1,0 @@
----
-spec: metrics.spec.md
----
-
-## Tasks
-
-- [x] Write spec files for metrics module
-- [x] Implement LOC counting by language
-- [x] Implement language detection from file extensions
-- [x] Implement comment/blank/code line classification
-- [x] Implement git churn analysis
-- [x] Implement test file detection and ratio
-- [x] Add JSON output mode
-- [x] Wire up CLI subcommand in main.rs
-- [x] Run verification suite
