--- specs/prs/tasks.md
@@ -1,13 +1,0 @@
----
-spec: prs.spec.md
----
-
-## Tasks
-
-- [x] Write spec files for prs module
-- [x] Implement `prs` list — styled PR listing with draft icons
-- [x] Implement `prs view` — detailed view with diff stats
-- [x] Add `--state`, `--limit`, `--json` flags
-- [x] Wire up CLI subcommands in main.rs
-- [x] Write unit tests
-- [x] Run verification suite
