+++ src/publish.rs
@@ -1,141 +1,7 @@
 use anyhow::{bail, Context, Result};
 use console::style;
-use dialoguer::{theme::ColorfulTheme, Confirm};
 use serde_json::json;
-use std::path::{Path, PathBuf};
-
-#[derive(Debug)]
-pub struct PublishOptions {
-    pub path: PathBuf,
-    pub org: Option<String>,
-    pub private: bool,
-    pub description: Option<String>,
-    pub yes: bool,
-}
-
-pub fn run(mut options: PublishOptions) -> Result<()> {
-    if crate::utils::is_non_interactive() {
-        options.yes = true;
-    }
-    let config = crate::config::Config::load()?;
-    let token = config.github_token().ok_or_else(|| {
-        anyhow::anyhow!(
-            "No GitHub token configured. Run: fledge config set github.token <your-token>"
-        )
-    })?;
-
-    let path = options
-        .path
-        .canonicalize()
-        .with_context(|| format!("Directory not found: {}", options.path.display()))?;
-
-    let manifest = validate_template(&path)?;
-
-    let repo_name = &manifest.template.name;
-    let description = options
-        .description
-        .as_deref()
-        .unwrap_or(&manifest.template.description);
-
-    let owner = match &options.org {
-        Some(org) => org.clone(),
-        None => get_authenticated_user(&token)?,
-    };
-
-    println!(
-        "{} Publishing {} as {}/{}",
-        style("➡️").cyan().bold(),
-        style(path.display()).dim(),
-        style(&owner).green(),
-        style(repo_name).green()
-    );
-
-    let sp = crate::spinner::Spinner::start("Checking repository:");
-    let repo_exists = check_repo_exists(&owner, repo_name, &token)?;
-    sp.finish();
-
-    if repo_exists {
-        if !options.yes {
-            crate::utils::require_interactive("yes")?;
-            let confirm = Confirm::with_theme(&ColorfulTheme::default())
-                .with_prompt(format!(
-                    "Repository {}/{} already exists. Push update?",
-                    owner, repo_name
-                ))
-                .default(false)
-                .interact()?;
-
-            if !confirm {
-                println!("{} Cancelled.", style("*").cyan().bold());
-                return Ok(());
-            }
-        }
-    } else {
-        let sp = crate::spinner::Spinner::start("Creating repository:");
-        create_github_repo(
-            repo_name,
-            description,
-            options.private,
-            options.org.as_deref(),
-            &token,
-        )?;
-        sp.finish();
-        println!(
-            "  {} Created repository {}/{}",
-            style("✅").green().bold(),
-            owner,
-            repo_name
-        );
-    }
-
-    let sp = crate::spinner::Spinner::start("Setting repository topics:");
-    set_repo_topics(&owner, repo_name, &token)?;
-    sp.finish();
-    println!(
-        "  {} Set {} topic",
-        style("✅").green().bold(),
-        style("fledge-template").cyan()
-    );
-
-    let sp = crate::spinner::Spinner::start("Pushing template files:");
-    push_directory(&path, &owner, repo_name, &token)?;
-    sp.finish();
-    println!("  {} Pushed template files", style("✅").green().bold());
-
-    println!(
-        "\n{} Published! Install with:\n\n  {}",
-        style("✅").green().bold(),
-        style(format!(
-            "fledge init <project-name> -t {}/{}",
-            owner, repo_name
-        ))
-        .cyan()
-    );
-
-    Ok(())
-}
-
-pub fn validate_template(path: &Path) -> Result<crate::templates::TemplateManifest> {
-    if !path.exists() {
-        bail!("Directory not found: {}", path.display());
-    }
-
-    let manifest_path = path.join("template.toml");
-    if !manifest_path.exists() {
-        bail!(
-            "No template.toml found in {}. Create one with: fledge create-template",
-            path.display()
-        );
-    }
-
-    let content = std::fs::read_to_string(&manifest_path)
-        .with_context(|| format!("reading {}", manifest_path.display()))?;
-
-    let manifest: crate::templates::TemplateManifest =
-        toml::from_str(&content).with_context(|| "Invalid template.toml")?;
-
-    Ok(manifest)
-}
+use std::path::Path;
 
 pub fn get_authenticated_user(token: &str) -> Result<String> {
     let text = ureq::get("https://api.github.com/user")
@@ -210,10 +76,6 @@ pub fn create_github_repo(
         }
         Err(e) => bail!("Failed to create repository: {}", e),
     }
-}
-
-pub fn set_repo_topics(owner: &str, repo: &str, token: &str) -> Result<()> {
-    set_repo_topic(owner, repo, "fledge-template", token)
 }
 
 pub fn set_repo_topic(owner: &str, repo: &str, topic: &str, token: &str) -> Result<()> {
@@ -360,104 +222,6 @@ pub fn run_git(dir: &Path, args: &[&str]) -> Result<()> {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    fn write_valid_manifest(dir: &Path) {
-        std::fs::write(
-            dir.join("template.toml"),
-            r#"[template]
-name = "test-template"
-description = "A test template"
-
-[files]
-render = ["**/*.md"]
-"#,
-        )
-        .unwrap();
-    }
-
-    #[test]
-    fn validate_valid_template_succeeds() {
-        let tmp = TempDir::new().unwrap();
-        write_valid_manifest(tmp.path());
-
-        let result = validate_template(tmp.path());
-        assert!(result.is_ok());
-        let manifest = result.unwrap();
-        assert_eq!(manifest.template.name, "test-template");
-        assert_eq!(manifest.template.description, "A test template");
-    }
-
-    #[test]
-    fn validate_missing_manifest_fails() {
-        let tmp = TempDir::new().unwrap();
-
-        let result = validate_template(tmp.path());
-        assert!(result.is_err());
-        assert!(result.unwrap_err().to_string().contains("No template.toml"));
-    }
-
-    #[test]
-    fn validate_invalid_manifest_fails() {
-        let tmp = TempDir::new().unwrap();
-        std::fs::write(tmp.path().join("template.toml"), "not valid toml {{{{").unwrap();
-
-        let result = validate_template(tmp.path());
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn validate_nonexistent_dir_fails() {
-        let result = validate_template(Path::new("/nonexistent/path/to/template"));
-        assert!(result.is_err());
-        assert!(result
-            .unwrap_err()
-            .to_string()
-            .contains("Directory not found"));
-    }
-
-    #[test]
-    fn validate_manifest_with_prompts() {
-        let tmp = TempDir::new().unwrap();
-        std::fs::write(
-            tmp.path().join("template.toml"),
-            r#"[template]
-name = "prompted"
-description = "Has prompts"
-
-[prompts.database]
-message = "Database engine"
-default = "sqlite"
-
-[files]
-render = ["**/*.rs"]
-"#,
-        )
-        .unwrap();
-
-        let result = validate_template(tmp.path());
-        assert!(result.is_ok());
-        let manifest = result.unwrap();
-        assert!(manifest.prompts.contains_key("database"));
-    }
-
-    #[test]
-    fn run_rejects_no_token() {
-        let tmp = TempDir::new().unwrap();
-        write_valid_manifest(tmp.path());
-
-        let options = PublishOptions {
-            path: tmp.path().to_path_buf(),
-            org: None,
-            private: false,
-            description: None,
-            yes: false,
-        };
-
-        let result = run(options);
-        assert!(result.is_err());
-    }
 
     #[test]
     fn topics_include_fledge_template() {
